Updated openstack/openstack Project: openstack-infra/project-config  4dd83d933f54a93c9ce9b6b92397806368c32afb
Remove check-requirements job for python-dracclient

This was never intended, and dracclient has never been in projects.txt.
This is also the only vendor support library in Ironic world that has it.

Finally, dracclient is branchless and is designed to support more than one
OpenStack release, which is harder with global requirements.

Change-Id: I2fd9777a16bc2cd35504ea002d22856eeeb5f8b9