Updated openstack/openstack Project: openstack-infra/storyboard  6453df36cdc52666b989ba1dfafde7cb31095bf2
configure tox to allow OS_TEST_TIMEOUT env var to pass through

By default tox does not pass environment variables to the processes it
runs. Add OS_TEST_TIMEOUT to the whitelist so that a developer running
the tests locally can update the timeout for the database tests that run
on a slow box.

Change-Id: I1300c0d8c0b8a0d5011fc5d2ac028f3ab2527593
Signed-off-by: Doug Hellmann <doug@doughellmann.com>