Updated openstack/openstack Project: openstack/cinder  e0e92fac9a064a2f527f0a6968e644d520c199c8
NetApp ONTAP: Deprecate 7-mode driver

For several years now, NetApp has been transitioning its
customers to a newer version of ONTAP that operates
in Cluster mode. We are now deprecating the 7-mode
Data ONTAP Cinder driver, in line with end of support for
the product.

The deprecation begins this release (Ocata) and the driver
will be removed in Queens. We will run the 7-mode CI until
stable/pike is tagged EOL; no new features will be added,
however, existing code may be meaningfully refactored to
keep the driver ready for removal.

DocImpact

Change-Id: I181e6be45c358c3d8b81b7a2cde2e1eb4ef50d92