Updated openstack/openstack Project: openstack/requirements  b170e22dc3e6ccac9e81d2c967a9a60c61762d72
Add xstatic-angular-ui-router

This library will be used by Horizon.

The underlying library (angular-ui ui-router) is actively maintained
and the xstatic packaging of it is an OpenStack project maintained
by the Horizon team.

It is good code.

The underlying library is not Python code, but Javascript, though the
xstatic wrapper code is Python code (and Python 3 compatible).

The license (MIT) is compatible.

It is not packaged for ubuntu or fedora yet, as the xstatic package
is too new.

The functionality of this library is not covered by any other existing
global requirement.

Change-Id: I87577ccdde018499df6684a73aea186309a915c5