Updated openstack/openstack Project: openstack/cinder  f8b0833c7332ed657d8e5e8c9a1622cfe206847b
storwize: get_host_from_connector optimization

Changed get_host_from_connector() to search the host entries
on the controller that match the connector's host property
first. This will reduce the search time for iSCSI and FC
attachments as the connectors host property is used as
input to create the host entry on the controller.

Added unit test cases.

Change-Id: Ib0a5ab483cf7110362884f31a39ca0df34ff39f3
Closes-Bug: #1632741