Updated openstack/openstack Project: openstack-infra/zuul  bafbc5b328a3584dc0d8f2c84e0306e84f083475
Ansible launcher: move AFS publisher into a module

The logic to rsync files into AFS is very complex, requiring
an rsync command for each of the pseudo-build-roots that are
produced by our docs jobs.  Rather than try to do this in ansible
YAML, move it into an ansible module where it is much simpler.

Change-Id: I4cab8003442734ed48c67e09ea8407ec69303d87