Updated openstack/openstack Project: openstack-infra/project-config  de25384af5463871683f1f0aeda659f85740e7d1
Fix scenario004 gates for newton

The puppet scenario004 does not work until after Newton.  Unfortunately
the job was improperly being added due to the ordering of the matcher.
The puppet4 blanket rule was overriding our disabling of scenario004.
This change fixes the order so that the scenario004 job should only be
active for master (Ocata) and beyond.

Change-Id: Ia8487d27317f5204bc7fc4d92dbe893fc9677aed