Updated openstack/openstack Project: openstack/diskimage-builder  587d14feed3cfdc6df9bf2d31898d5c81a377710
dhcp-all-interfaces: support Centos/RHEL 6

On Centos and RHEL 6 the init system is upsart but but networking is using
sysv compatabiliy and a code path the handle this situation.

We can't use DISTRO_NAME because the centos-minimal element sets it to
centos for CentOS 7 but the centos element sets it to centos for CentOS 6.

Change-Id: Ib8e33ed78b3d6a5737eb7449bccef2d33f72b131
Closes-Bug: #1638527