Updated openstack/openstack Project: openstack/neutron  475241834dae0d9e252e693dfca35fbff1513a57
Update metadata proxy when subnet add/delete

With current code, if first subnet of the network is an ipv6 subnet,
the metadata proxy will not be spawned. If user then adds ipv4 subnet
with dhcp enabled, the metadata proxy will still not be spawned. As a
result, the metadata service will not be available for the network.

This patch will kill/spawn metadata proxy,  when subnet add/delete.
So, even if the first subnet of the network is not an ipv4 subnet with
dhcp enabled, the metadata proxy can still be spawned if network has
subnets need metadata proxy.

Closes-bug: #1556991

Change-Id: I0b45af8f2b756732f45c13d7e2dbcd30653cc026