Updated openstack/openstack Project: openstack/governance  9b1912ad91697ff758cc901910d5687bc5ad8985
Create a project tag for zero-impact upgrades

The asserts:supports-zero-downtime-upgrade tag allows for a performance
penalty during the upgrade process, such as delayed responses for
service APIs. This tag raises the bar further by eliminating that
allowance.

As a result of this tag, services are expected to be able to support
upgrading:

* in a rolling fashion (running multiple versions of a service
concurrently),

* without incurring any API downtime (API responses must return without
failure),

* and without incurring any performance penalty (API responses must return
without a perceivable delay from normal operations).

With this patch, the upgrade-related tags would then include:

1. asserts:supports-upgrade
2. asserts:supports-rolling-upgrade
3. asserts:supports-zero-downtime-upgrade
4. asserts:supports-zero-impact-upgrade

Change-Id: If73d936977d8fcab02d1f7be42e00b304391f4c6