Updated openstack/openstack Project: openstack/glance  09ca4452224897753074b78a8d729e4ab500d271
Added overwrite warning for db_export_metadefs.

The bug in [1] proposed to always require a directory option for
glance-manage db_export_metadefs command. The rationale being that
silently overwriting the default path /etc/glance/metadefs can
be undesired. However, the feedback was that this command should
stay silent in case it's called from cron jobs or automated tasks,
and that it's better to simply log a warning and proceed, as well
as update the documentation with a cautionary note. This patch
addresses that request.

[1] https://bugs.launchpad.net/glance/+bug/1367011

Change-Id: Ie8cc4bb3769a6347fd25e2235a72c6358af70d42
Closes-Bug: #1367011