Updated openstack/openstack Project: openstack-infra/gear  a74b8d50b8b8c14c97fc528c903fa59a8f555b58
Re-enable flake8

The line in tox.ini "select=H231" did not add H231 to the list of flake8
checks. It made us only run H231.

Remove the H231 line and then fix the flake8 issues. The most
substantial of these was the shadowing of the queue module with local
variables named queue. Made that one import queue as queue_mod because
changing the variable names in the code was yuckier to look at.

Change-Id: I4a26a20889132f7f4525b17392088dda6bd5bbd2