Updated openstack/openstack Project: openstack/python-cinderclient  b3de9894373e5113ba8d10484937967cf04e33de
Update --endpoint-type dest to os_endpoint_type

Currently the dest of the deprecated --endpoint-type arg
is 'endpoint_type'. However, the code only uses 'os_endpoint_type'
so right now it isn't being used at all. This patch updates
the dest to be 'os_endpoint_type' so that the variable can
be assigned correctly.

Change-Id: Ie8592d12bcf95e4ccff571e831440b18f5acfd40
Closes-bug: #1454436