Updated openstack/openstack Project: openstack/neutron  c4c3fbdb5e5e0f410458833f6e737df91ed60b70
Don't pass config object to send_ip_addr_adv_notif()

Previously, the function consumed a config object and picked only a
single option send_arp_for_ha from that object. This patch replaces the
config object with count itself in order to be able to send gratuitous
ARPs not relying on a single config option.

As per http://codesearch.openstack.org, this function is not called in
any other stadium project. The send_arp_for_ha option is deprecated as
for Ocata.

Related-bug: 1639315

Change-Id: I6efc539a969bebb4fd969e1648a787bf7dc9cd0b