Updated openstack/openstack Project: openstack/nova  61b50feadeefd547f8452188f5d4a8a4219caef4
Fix a TypeError in notification_sample_base.py

In notification_sample_base, a possible error is found and fixed, as

    extra_params['return_reservation_id'] = True

is a valid code only when extra_params is not None.
However, this is not checked. So when we try to boot a server
without extra parameters, a TypeError is raised:

    TypeError: 'NoneType' object does not support item assignment

In this patchset, the possible bug is fixed.

Change-Id: I80e2286b231da39b8c1e4225145422ac48daee68