Updated openstack/openstack Project: openstack/oslo.rootwrap  c2f337b9b6ddf35b704d28f4a5facbbe1d37225d
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ib1a145799101da6196445e8ed78ba70863df30e2