Updated openstack/openstack Project: openstack/oslo.context  522b2110334be5fb9a9a9b92b96cb10b0c2db4eb
Provide a way to deprecate policy values

If we want to move all the services over to a standard policy
enforcement dict we need a way to start notifying deployers when the
policy enforcement files are using deprecated values.

Instead of returning a dictionary return an object that acts like a
dictionary but emits a DeprecationWarning whenever a deprecated policy
enforcement value is read from it.

Change-Id: I4b2fda188bbccfd491556cc5631e5c4a76314492