Updated openstack/openstack Project: openstack/keystone  e70631eddad2b26a3e5185fec9c379e03b8acb70
Add healthcheck middleware to pipelines

This introduces the oslo healt check middleware
http://docs.openstack.org/developer/oslo.middleware/healthcheck_plugins.html
into the pipelines. This middleware is useful for load balancers and
http servers, which can use it to validate that the keystone services are
operational. This middleware is being used in other services such as
glance and magnum. This patch provides it for keystone, in an effort to
spread the usage across all major projects.

This is one less item that operators will have to patch locally.

DocImpact

Change-Id: I19e4fc8f6c6a227068ba7191c1e9c453fc08f061