Updated openstack/openstack Project: openstack-infra/project-config  5ac30c8223ba08f53bb3078e4b04d8170a6285b1
Stop uploading images to demo european clouds

Image uploads are a thing we struggle with getting done reliably. Reduce
the overhead by removing the BCN keynote demo clouds as we are not using
them in production currently. Hopefully we can readd these clouds in the
future but for now simplify nodepool and remove them.

Change-Id: I0e099bdf9b4e95efd9837b3acb99292239c1a281