Updated openstack/openstack Project: openstack/nova  72d0c3a484a17fa67ec2827377c15c25e8b8efda
libvirt: remove py26 compat code in "get_console_output"

The "get_console_output" interface had code for python2.6 compatibility.
As we need at least python2.7, we can drop that code, which makes the
method (a little) easier to read.
The unit tests don't need any change, as the expected result is the same.

Change-Id: Ic3c22b96937a836dbc50b7ce3b2ba1a864450180