Updated openstack/openstack Project: openstack-infra/project-config  85d5a2dc4d18d1513403edbce5b75445b4e11474
Restrict nova-net grenade jobs to <= Newton

Nova is going to make nova-network fail to start except
in cells v1 configurations, so we need to restrict all
non-neutron grenade jobs to run on <= Newton stable branches.

This is mostly targeted at the jobs defined in the
grenade-nova-net group, and the -forward and -multinode
grenade jobs.

Change-Id: I62348f39b58e188b85d18918cfd302fa2d006907