Updated openstack/openstack Project: openstack-infra/project-config  3502dab2c82ea2889e3b2c4ed8c737cb55eca228
Restrict gate-tempest-dsvm-multinode-full-ubuntu-xenial to Newton

Just like with the gate-tempest-dsvm-full-ubuntu-xenial job,
the multinode variant of that nova-network job should be restricted
to stable/newton since we're going to switch devstack-gate to run
with neutron by default in master (ocata), and nova-network won't
work except in cells v1 configurations.

This just updates the regex to accont for the multinode nova-net
job in addition to the single node job.

Change-Id: Ie061c3c18a1b7ef23d6bff8bb6f705fd5ec65ca5