Updated openstack/openstack Project: openstack/keystone  dacd57f972d9c1a8f8b268067c2c763fb69d5cc3
Request cache should not update context

The request context is kind of important. It has info like the
request cache and auth info for the request.

The request cache should not be modifying the request context. The
code here makes it look like the request cache needs to update the
thread's request context but it actually never does change it since
even if it creates a new RequestContext the new instance becomes
the new thread request context.

Change-Id: Ib824648fe7cb6b3614e03f7d2570fbbfc0c3a88d