Updated openstack/openstack Project: openstack/nova  b2fd340e46ea895d01b9bb80f8a3ed14fb5843ca
libvirt: Fix incorrect libvirt library patching in tests

In several places were were incorrectly globally patching modules
which load the libvirt library in tests. A number of tests relied on
this patching, but did not do it themselves. They happened to work
when running all tests due to test ordering, but failed when running a
subset of tests.

Change-Id: Ic03a50ef702490abad7482213a8f118318422064