Updated openstack/openstack Project: openstack/horizon  a74352df88ce4f212b570410c684418e1a652933
Make "Detach Volume" in Instances page translatable

Due to tricky behavior of gettext, if we have both normal
gettext string and plural gettext string for a same string,
the normal gettext string is not translated.
(The detail is described in the bug report. See comment #5.)

As a workaround, this commit adds a context information to
the plural gettext string in volumes.volumes.tables.DetachVolume
so that we can have separate entries for a normal entry and
an entry with plural form support. By doing so, gettext can
find an expected translated string for a normal string.

After this change, we will see the following entries
for "Detach Volume" in the generated PO file.
----
msgid "Detach Volume"
msgstr ""

msgctxt "Action to perform (the volume is currently attached)"
msgid "Detach Volume"
msgid_plural "Detach Volumes"
msgstr[0] ""
msgstr[1] ""
----

Note that the message "Detaching Volume" for the post action
in the table is also updated to use npgettext_lazy.
This and "Detach Volume" are a pair, so it looks better to
use a same style.

Change-Id: Idf772265d3e0e1d2e74b19755a467e1433233f64
Closes-Bug: #1626312