Updated openstack/openstack Project: openstack-infra/tripleo-ci  f7c144ea3a4cd86ce3cf7eafb154bd3ddf678d1c
Redirect image build stderr to stdout

Image build output is redirected to a file to avoid cluttering up
the console log, but stderr is not redirected so it ends up in a
weird place in the console log that makes it hard to track down
errors.  This change redirects image build stderr to stdout so it
will also be captured in the file.  Note that this is already done
for the undercloud install where the same thing happens.

Change-Id: I3d1a7d95eaac1fcfaa75cbe674085f3d9aca1410