Updated openstack/openstack Project: openstack/tripleo-heat-templates  1e11964372ee2d719c2f951c1f3820f35f3e225a
Ensure heat-domain hiera is in nodes that contain keystone

The commit that this depends on only works if heat is deployed in the
same node as keystone. Once we deploy them in different nodes, keystone
won't be able to retrieve the appropriate hieradata. This fixes that by
setting the appropriate hieradata to be deployed on the keystone service
by the heat profiles.

Change-Id: I1f08db68a14486526879d1a5a1ff78cb17686924
Depends-On: I7d42d04ef0c53dc1e62d684d8edacfed9fd28fbe