Updated openstack/openstack Project: openstack-infra/infra-specs  5a3abf970c3bb52fb9fe0e175c0f37cdd37c3c1d
Mark openstackci spec as complete

The openstackci spec has been completed for a while.
The last portion for kibana/elk was listed as optional,
but not needed for third party ci. Future work to make this
and other puppet modules re-consumable by others can continue
without this spec or as a priority effort.

Change-Id: I79950ae8008df571145fe207cf395f5a3de8df6f