Updated openstack/openstack Project: openstack/nova  f8e35e5bf969d44fd0e127b01cb62f95ff760071
neutron: actually populate list in populate_security_groups

We're going to store the list of requested security group objects
in the request spec so the neutron security group driver code
needs to actually populate that list now for use in the compute API.

As the nova-network and neutron security group API code is basically
the same for populating that list, the backend-specific implementations
are removed and pulled up into the base security group driver API class.

Since we don't want to actually try and persist this list with the
instance object in the nova database we need to add a check in the
compute API to store an empty list, like the old behavior, when
creating the instance. A later change will use the populated list
to store in the RequestSpec.

Part of blueprint cells-scheduling-interaction

Change-Id: I6139eb92f534f010565fde31661382d33463ae58