Updated openstack/openstack Project: openstack-infra/project-config  9dd784fe07354cf74619cbf079937fed8fd27f28
Switch ironic-ipa jobs to Xenial and make ipmitool jobs voting

This patch is switching some of the ironic-ipa jobs to Ubuntu Xenial
for the master branch and also switching the ipmitool jobs (which runs
on Xenial) to voting.

Xenial jobs (non-votin) were also added to nova, neutron and tempest
check gates to replace the Trusty ones due the policy to not run Trusty
on newton or master branches.

Change-Id: I7812d23c11ed92859cb59d32ddaac770ea610449