Updated openstack/openstack Project: openstack/cinder  b15a6b3e6fd4dcdab5a3bd6246db081c52b497e0
DB: Optimize volume_update method

Our current volume_update method reads the volume from the DB, joining
the volume table with 4 or 5 other tables, and then do the update using
SQLAlchemy ORM engine before returning updated ORM object.

But the truth is that we don't really need to retrieve the volume again
in any of existing cases, since the caller either doesn't care about
current value in the DB -for example when setting status to error- or it
already has that information -composing it with the current value it has
and the updated fields it's passing- and doesn't need to reload it.

This additional joined read from the DB on every volume update is
creating unnecessary load on the DB and reducing our network capacity
and in other resources we are not retrieving it either.

This patch modifies the way we do updates so we don't read current DB
values.

Change-Id: I92e76eb8eb64ac96995b29017c7683a564dc5e0e