Updated openstack/openstack Project: openstack/nova  c4b1cdfc368f9035da61c2901784c4515672ff00
Add flavor extra_spec info link to api_ref

In the current api-ref, we have illustrated
about how to add extra specs for flavors,
but it was very simple and we have a large
number of built-in extra specs that can be
used to archive different purpose listed here:
http://docs.openstack.org/admin-guide/compute-flavors.html

The reader will still be confused about how
to correctly use extra specs only read the
api-ref. We could first simply add the above
link to the flavor extra spec section in
api-ref to let readers know where to find
information and come up with better solution
later.

Change-Id: Ie720973783f153d3954be9f4bd97e7f9bf788c6b
Closes-bug: #1640654