Updated openstack/openstack Project: openstack/python-cinderclient  c99fb20067fc0b6927591a540c2f5d6cec6866bf
Mask passwords when logging HTTP req/resp bodies

The very specific 'password' is being masked when logging
requests but not when logging response bodies.

This change fixes the response logging to mask passwords and
also makes the request logging more robust as it was just
checking for 'password' but the mask_password method handles
much more than that.

Change-Id: Id8bf583ecdf60eafb50fd69d6a19180ea97bd92c
Closes-Bug: #1640269