Updated openstack/openstack Project: openstack/nova  8a65b3d83c7b1809518b4e743ef0629e3899f3c7
Catch all local/catch-all addresses for IPv6

_check_graphics_addresses_can_live_migrate() checks to see if an address
is local or catch all by doing a match against

            LOCAL_ADDRS = ('0.0.0.0', '127.0.0.1', '::', '::1')

In IPv6, due to zero compression, there are many different valid ways
to represent :: and ::1 (since leading zeros don't need to be suppressed).

Because of this, the current code will fail to match against something
like "::0" even though it's a perfectly valid way to represent the
same address.

The simple fix for this is to convert the address(es) in the config file
to canonical form before matching against LOCAL_ADDRS.

Change-Id: I13dd45fac489e722747f74695567a40a3a201652
Closes-bug: #1634949