Updated openstack/openstack Project: openstack/swift  2b116cffd3943404029749bbc6f1f8d25b40795e
Include debug message for rsync tempfiles

The warning message for rsync tempfiles was removed in the related
change.  However, because our regex match might result in a false
positive maybe it's still useful to log a debug message.  Instead of
silently ignoring rsync tempfiles, when running in debug we note the
file and how we classified it - but still no warning will occur.

I also consolidate our use of the regex for rsync tempfiles into the
diskfile module and move the negative test for the warning logger
message a little next to the positive test.

Change-Id: Idb2a1a76aa275c9c2e9bad8aceea913b8f5b1c71
Related-Change: I5a5d6e24710e4880776b32edcbc07021acf77676