Updated openstack/openstack Project: openstack/neutron  f46cba17f0a995dd0afab240bc2c64beb09c3639
Truncate IPDevice's name to interface max size

An interface's name is limited to INTERFACE_MAX_SIZE. In some cases, IP
utils truncate this name automatically. In Ubuntu Xenial, ip link
returns an error instead of this behaviour.

This change truncates the device's name upon read (The full name is
still stored on the object). This way, the code does not rely on IP
utils to do the truncation. This solves the error in the test.

Change-Id: I41b226e4b52239b861e2b6e806296025502b9b60
Closes-Bug: 1637993