Updated openstack/openstack Project: openstack/nova  4e0ba577ef62dd7f9a6d42737be2b86b15a5a286
Pass MTU into os-vif Network object

This passes the MTU from the network_info into the
constructed os-vif object so it can be used by the
os-vif plugging libraries to set the correct MTU.

Now with os-vif >=1.3.0 in the global requirements,
os-vif has correct support for MTUs on the os-vif
objects so we can use them correctly.

The follow-up patch removes the MTU hacks that were
added last minute for the related bug.

Related-Bug: #1623876
Change-Id: Iae0d83d22511f35f00ed6d8fde03f8c32dbedb26