Updated openstack/openstack Project: openstack/cinder  a259d85a5715bc038042e729b9381f57ca14632a
Set EVENTLET_NO_GREENDNS for IPv6 and dnspython compat

If dnspython is present it causes socket.getaddrinfo() to be
monkeypatched wiht code that does not work with IPv6. To get
around this, the EVENTLET_NO_GREENDNS environment variable
must be set before importing eventlet.

This is resolved in the eventlet upstream code [1], but has not
yet been released. Once our requirements have been raised past
the 0.19.0 release with a version that includes this change we
can remove this workaround.

[1] https://github.com/eventlet/eventlet/commit/52b09becacd23f384cf69ae37d70c893c43e3b13

Change-Id: Id9fe265d67f6e9ea5090bebcacae4a7a9150c5c2
Closes-bug: #1471049