Updated openstack/openstack Project: openstack/neutron  b51c5635ab63ad720c3ea0f07209c5d0599c2c0d
Changing arping command execute to accept 1 as extra OK code

Connecting any network to a router causes l3-agent to send out
gratuitous ARP packets that do not require a response. On Ubuntu 14.04
sometimes this command returns 1, which causes it to falsely report
arping command failure. This is now added to the extra OK codes.

Change-Id: I6299e936d4e4e7fd7497d23224e7abf605a99c15
Closes-bug: #1578842